requirement.txt
requests
chirpstack_api
mysql-connector-python